Adds filename working notes
Given that the Ops Manager uploads are being handled by v2.3.0 of the
nginx-upload-module, and given that this code sure does look like it's
responsible for creating the name of the file that is used by the upload
module to store the module, this looks like a decent place to try the
"Use the presence of a file with the same name as the file that's being
 uploaded as an indicator that an upload is in progress and we should
 abort subsequent uploads" strategy.

The idea (as sort-of described in the code comments I've added) is that
we'll name the upload temporary file the same as the filename in the
request information. Part of the initialization of the upload process is
creating a temporary file in which to put the bits to pass along to the
Ops Manager. I EXPECT that we can reconfigure the call that creates this
temporary file to fail in a way that makes the request fail with a
reasonable HTTP response code when that file already exists.

(NB: I HOPE but do not know that the _real_ filename is is in the
 `u->request` argument to the `ngx_http_upload_start_handler` (I ALSO
 HOPE that the name in the request is the same as the name of the file
 that Ops Manager is told...  if it isn't, then we'll need to find out
 how Ops Manager knows to get the right filename.)

ALSO NOTE that we're relying on the "delete the temporary file once the
upload is actually done" logic to be correct... and to never leave stray
files around. We don't KNOW that it's correct, but it _looks_ like it's
using `nginx`-provided mechanisms to get its work done... so hopefully
the only way it can fail is if `nginx` -itself- crashes hard, in which
case all bets are off.

ANYWAY.

This commit adds notes from today's research. Useful documentation is
pretty hard to come by (probably because the nginx folks really, really
want you to give them money for paid support) but some somewhat useful
docs links are as follows:

<https://www.nginx.com/resources/wiki/extending/api/utility/#functions>
(for some of the various memory manipulation functions, like
 `ngx_memcpy` and `ngx_sprintf`)

<https://www.nginx.com/resources/wiki/extending/api/alloc/>
(for the pool memory manipulation functions (like `ngx_palloc`))

I've (unfortunately) found that the best strategy to find these docs is
to put the name of the `ngx_XXX` function of interest into Google and
clicking the link to the nginx.com docs... the built-in search on the
docs page DOES NOT WORK.

[#184826410]