More notes about extracting form-data filename
After a bit of research, it seems very likely that we ALREADY have the
form-data filename in the `u->file_name` member. If the module is
handling a 'multipart/form-data' request, then it looks like the code
populates that with the original filename. More details in the comments
added by this commit.