Correctly copies over filename chars
This commit painstakingly copies over the characters in the temporary
directory path and the filename calculated from formdata. It turns out
that the filename in the formdata is not reliably null-terminated (which
means that it's NOT null-terminated), so we have to memcpy over the bits
into the filename.

Anyway. With this change, we reliably return 400s when the file already
exists on disk, preventing simultaneous uploads. Our primary CLI driver
gives up after two or three tries, so it fails fast... thankfully.

TODO: RTFM and see if there are specialized "copy string data into a
buffer" and/or "create a path out of these components" functions in the
nginx standard library. All this manual memcpy stuff surely isn't the
best way to do this.