Fail early if asked to do partial content
Because the feature uses the presence of a file on disk to determine
whether an upload is in progress or not, this probably makes it
incompatible with partial-content... which would require us to add
additional data to an existing file.

So, fail early if this setting is on and the upload is partial-content.

This commit also adds "Turn the feature on and test existing behavior"
test files. It very conspicuously avoids testing HTTP/2, as those tests
don't pass in upstream's code... without these changes.

NOTE: read the FIXME comment added by this commit for info about a
      newly-added failing Test.